feat: users can only see and manage own tasks

+++ src/routers/task_routers.py
@@ -3,7 +3,7 @@ from sqlalchemy.orm import Session
 from src.database.db import get_db
 from src.models.task import Task
 from src.schemas.task_schema import TaskCreate, TaskResponse, TaskUpdate
-from src.utils.helpers import get_task_or_404
+from src.utils.helpers import get_task_or_404, verify_task_owner
 from src.utils.dependencies import get_current_user
 from src.models.user import User
 task_router = APIRouter(prefix="/tasks", tags=["Tasks"])
@@ -27,7 +27,7 @@ def get_tasks(
     limit: int = 10 ,
     db: Session = Depends(get_db),
     current_user: User = Depends(get_current_user)):
-   query = db.query(Task)
+   query = db.query(Task).filter(Task.user_id == current_user.id)
 
    if completed is not None:
        query = query.filter(Task.completed == completed)
@@ -40,12 +40,15 @@ def get_tasks(
 #-----------GetTask--------------------
 @task_router.get("/{task_id}", response_model=TaskResponse)
 def get_task(task_id: int, db: Session = Depends(get_db), current_user: User = Depends(get_current_user)):
-    return get_task_or_404(task_id, db)
+    task = get_task_or_404(task_id, db)
+    verify_task_owner(task, current_user)
+    return task
 
 #-----------UpdateTask--------------------
 @task_router.put("/{task_id}", response_model=TaskResponse)
 def update_task(task_id: int, task_data: TaskUpdate, db: Session = Depends(get_db), current_user: User = Depends(get_current_user)):
     task = get_task_or_404(task_id, db)
+    verify_task_owner(task, current_user)
     for field, value in task_data.model_dump(exclude_unset=True).items():
         setattr(task, field, value)
     db.commit()
@@ -56,6 +59,7 @@ def update_task(task_id: int, task_data: TaskUpdate, db: Session = Depends(get_d
 @task_router.delete("/{task_id}")
 def delete_task(task_id: int, db: Session= Depends(get_db), current_user: User = Depends(get_current_user)):
     task = get_task_or_404(task_id, db)
+    verify_task_owner(task, current_user)
     db.delete(task)
     db.commit()
     return {"message": "Task deleted"}
